feat: edit and delete existing Creators

+++ src/main/java/com/tuka/comiccharacters/dao/Dao.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface Dao<T> {
     void save(T entity);
 
-    T findById(int id);
+    T findById(Long id);
 
     List<T> findAll();
